+++ package_managers/debian/main.py
@@ -12,6 +12,10 @@ from core.models import URL, LegacyDependency, Package, PackageURL
 from core.scheduler import Scheduler
 from core.structs import Cache, URLKey
 from package_managers.debian.db import DebianDB
+from package_managers.debian.debian_sources import (
+    build_package_to_source_mapping,
+    enrich_package_with_source,
+)
 from package_managers.debian.diff import DebianDiff
 from package_managers.debian.parser import DebianData, DebianParser
 
@@ -22,13 +26,7 @@ BATCH_SIZE = 500
 
 
 def fetch(config: Config) -> tuple[GZipFetcher, GZipFetcher]:
-    should_fetch = config.exec_config.fetch
-    if not should_fetch:
-        logger.log("Fetching disabled, skipping fetch")
-        return None, None
-
-    logger.debug("Starting Debian package fetch")
-
+    """Fetches the Debian packages & sources manifest files"""
     package_source = config.pm_config.source[0]
     sources_source = config.pm_config.source[1]
     no_cache = config.exec_config.no_cache
@@ -39,128 +37,45 @@ def fetch(config: Config) -> tuple[GZipFetcher, GZipFetcher]:
         source=package_source,
         no_cache=no_cache,
         test=test,
-        file_path="",
+        file_path="",  # will autosave in data/debian/latest
         file_name="packages",
     )
-    package_files = package_fetcher.fetch()
-    logger.log(f"Fetched {len(package_files)} package files")
-    package_fetcher.write(package_files)
 
     sources_fetcher = GZipFetcher(
         name="debian",
         source=sources_source,
         no_cache=no_cache,
         test=test,
-        file_path="",  # just save it in data/debian/latest/
+        file_path="",  # will autosave in data/debian/latest
         file_name="sources",
     )
-    sources_files = sources_fetcher.fetch()
-    logger.log(f"Fetched {len(sources_files)} sources files")
-    sources_fetcher.write(sources_files)
+
+    # Fetch
+    should_fetch = config.exec_config.fetch
+    if should_fetch:
+        package_files = package_fetcher.fetch()
+        package_fetcher.write(package_files)
+        logger.log(f"Fetched {len(package_files)} package files")
+
+        sources_files = sources_fetcher.fetch()
+        sources_fetcher.write(sources_files)
+        logger.log(f"Fetched {len(sources_files)} sources files")
 
     return package_fetcher, sources_fetcher
-
-
-def build_package_to_source_mapping(sources_file_path: str) -> dict[str, DebianData]:
-    """
-    Build a mapping from binary package names to their source information.
-
-    Args:
-        sources_file_path: Path to the sources file
-        test: Whether to limit parsing for testing
-
-    Returns:
-        Dictionary mapping binary package names to source DebianData objects
-    """
-    logger.debug("Building package-to-source mapping")
-
-    # Parse sources file
-    with open(sources_file_path) as f:
-        sources_content = f.read()
-    sources_parser = DebianParser(sources_content)
-
-    # Build mapping: binary_package_name -> source_debian_data
-    package_to_source: dict[str, DebianData] = {}
-
-    for source_data in sources_parser.parse():
-        # Each source may produce multiple binary packages
-        if source_data.binary:
-            # Source has explicit binary list
-            for binary_name in source_data.binary:
-                binary_name = binary_name.strip()
-                if binary_name:
-                    package_to_source[binary_name] = source_data
-        else:
-            # No explicit binary list, assume source name == binary name
-            if source_data.package:
-                package_to_source[source_data.package] = source_data
-
-    logger.log(
-        f"Built mapping for {len(package_to_source)} binary packages from sources"
-    )
-    return package_to_source
-
-
-def enrich_package_with_source(
-    package_data: DebianData, source_mapping: dict[str, DebianData]
-) -> DebianData:
-    """
-    Enrich a package with its corresponding source information.
-
-    Args:
-        package_data: The package data from packages file
-        source_mapping: Mapping from package names to source data
-
-    Returns:
-        Enriched DebianData with both package and source information
-    """
-    # Start with the package data
-    enriched = package_data
-
-    # Determine source name
-    binary_name = package_data.package
-
-    # Look up source information
-    if binary_name in source_mapping:
-        source_data = source_mapping[binary_name]
-
-        # Enrich package with source information
-        # Only add source fields that aren't already populated
-        if not enriched.vcs_browser and source_data.vcs_browser:
-            enriched.vcs_browser = source_data.vcs_browser
-        if not enriched.vcs_git and source_data.vcs_git:
-            enriched.vcs_git = source_data.vcs_git
-        if not enriched.directory and source_data.directory:
-            enriched.directory = source_data.directory
-        if not enriched.build_depends and source_data.build_depends:
-            enriched.build_depends = source_data.build_depends
-        if not enriched.homepage and source_data.homepage:
-            enriched.homepage = source_data.homepage
-
-    else:
-        # Log warning for missing source
-        source_name = package_data.source or package_data.package
-        logger.warn(
-            f"Binary '{binary_name}' of source '{source_name}' was not found in sources file"
-        )
-
-    return enriched
 
 
 def run_pipeline(config: Config, db: DebianDB):
     """A diff-based approach to loading debian data into CHAI"""
 
     package_fetcher, sources_fetcher = fetch(config)
-
-    input_dir = "data/debian/latest"
+    input_dir = f"{sources_fetcher.output}/latest"
 
     # Build package-to-source mapping first
     sources_file_path = os.path.join(input_dir, "sources")
     if not os.path.exists(sources_file_path):
         logger.error(f"Sources file not found at {sources_file_path}")
         return
-
-    source_mapping = build_package_to_source_mapping(sources_file_path)
+    source_mapping = build_package_to_source_mapping(sources_file_path, logger)
 
     # Parse packages file
     packages_file_path = os.path.join(input_dir, "packages")
@@ -175,7 +90,9 @@ def run_pipeline(config: Config, db: DebianDB):
     # Process each package and enrich with source information
     enriched_packages: list[DebianData] = []
     for package_data in packages_parser.parse():
-        enriched_package = enrich_package_with_source(package_data, source_mapping)
+        enriched_package = enrich_package_with_source(
+            package_data, source_mapping, logger
+        )
         enriched_packages.append(enriched_package)
 
     logger.log(f"Processed {len(enriched_packages)} enriched packages")
@@ -275,11 +192,9 @@ def run_pipeline(config: Config, db: DebianDB):
             sources_fetcher.cleanup()
 
 
-def main():
+def main(config: Config, db: DebianDB):
     logger.log("Initializing Debian package manager")
-    config = Config(PackageManager.DEBIAN)
-    db = DebianDB("debian_db", config)
-    logger.debug(f"Using config: {config}")
+    logger.debug(f"Config: {config}")
 
     if SCHEDULER_ENABLED:
         logger.log("Scheduler enabled. Starting schedule.")
@@ -303,4 +218,6 @@ def main():
 
 
 if __name__ == "__main__":
-    main()
+    config = Config(PackageManager.DEBIAN)
+    db = DebianDB("debian_db", config)
+    main(config, db)
